+++ pac4j-core/src/main/java/org/pac4j/core/profile/factory/ProfileManagerFactory2.java
@@ -1,0 +1,16 @@
+package org.pac4j.core.profile.factory;
+
+import org.pac4j.core.context.WebContext;
+import org.pac4j.core.context.session.SessionStore;
+import org.pac4j.core.profile.ProfileManager;
+
+import java.util.function.BiFunction;
+
+/**
+ * A {@link ProfileManager} factory based on the {@link WebContext} and on the {@link SessionStore}.
+ *
+ * @author Jerome LELEU
+ * @since 4.0.0
+ */
+public interface ProfileManagerFactory2 extends BiFunction<WebContext, SessionStore, ProfileManager> {
+}